Phase2/requirements.txt
virtualenv
Pillow
exif
gpsphoto
piexif
exifread
